promjenio price u double i napravio nasljedivanje

+++ caffeApp/Consumable.cs
@@ -9,9 +9,9 @@ namespace VubCaffe
     public abstract class Consumable : IProduct
     {
         public string Name { get; protected set; }
-        public decimal Price { get; protected set; }
+        public double Price { get; protected set; }
 
-        public Consumable(string name, decimal price)
+        public Consumable(string name, double price)
         {
             Name = name;
             Price = price;
